#26 refactor: userRepository 패키지 이동
