+++ src/main/java/com/susanghan_guys/server/global/security/CustomUserDetailService.java
@@ -3,7 +3,7 @@ package com.susanghan_guys.server.global.security;
 import com.susanghan_guys.server.global.common.code.ErrorCode;
 import com.susanghan_guys.server.global.exception.BusinessException;
 import com.susanghan_guys.server.user.domain.User;
-import com.susanghan_guys.server.user.infrastructure.UserRepository;
+import com.susanghan_guys.server.user.infrastructure.persistence.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
